add option for continue downloading, retry until wget returns zero(seems wget doesn't retry automatically under some circumstances even with the -t 0 option)

+++ flvcd.pl
@@ -1,11 +1,14 @@
 #!/usr/bin/perl -w
 use strict;
 use http;
-my ($f, $i, $d) = ("", 0, 0);
+my ($f, $i, $d, $c) = ("", 0, 0, 0);
 while ($i < @ARGV) {
 	if ($ARGV[$i] eq "-f") {
 		$f = $ARGV[$i+1];
 		$i += 2;
+		next;
+	} elsif ($ARGV[$i] eq "-c") {
+		$c = 1, $i++;
 		next;
 	}
 	$d=1, $i++, next if ($ARGV[$i] eq "-d");
@@ -20,14 +23,16 @@ while ($i < @ARGV) {
 		if ($ans[$i] =~ /自动切割的<font color=red>(\d+)/) {
 			$cnt = $1;
 			for (0..$cnt-1) {
+				next if $c && -e "$_.flv";
 				$ans[$i + $_] =~ /href="(.*)" target=/;
 				qx|wget -U Xeslaro --no-dns-cache --connect-timeout=10 -t 0 -O $_.flv "$1"|;
+				qx|wget -U Xeslaro --no-dns-cache --connect-timeout=10 -t 0 -O $_.flv "$1"| while ($?);
 			}
 			last;
 		}
 		$i++;
 	}
-	$cnt--;
+	$c=0, $cnt--;
 	qx|mencoder -ovc copy -oac lavc -lavcopts acodec=ac3 -of lavf -lavfopts format=matroska -o "$name".mkv {0..$cnt}.flv|;
 	qx|rm {0..$cnt}.flv| if ($d);
 }
